requirements.txt
# PDF 텍스트 추출
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# 텍스트 청킹
tiktoken>=0.5.0

# 임베딩 및 벡터 DB
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.4

# 시각화
plotly>=5.18.0
scikit-learn>=1.3.0
numpy>=1.24.0

# LLM
openai>=1.0.0
groq>=0.4.0

# 웹 데모 (선택)
streamlit>=1.28.0
gradio>=6.0.0

# 유틸리티
python-dotenv>=1.0.0
tqdm>=4.66.0


